+++ src/domain/builder/filetype/markdown.builder.ts
@@ -1,7 +1,8 @@
 import { parse } from "marked";
 
-import { IBuilder, IFile, ICache } from "../../../interfaces";
-import { htmlBuilder } from "./html.builder";
+import { IBuilder, IFile, IBuilderCache } from "../../../interfaces";
+import { buildHtml } from "./html.builder";
+import { ICache } from "../../../lib/cache";
 
 const parseMarkdown = async (markdown: string) => {
   return new Promise<string>((resolve, rejects) => {
@@ -12,16 +13,15 @@ const parseMarkdown = async (markdown: string) => {
   });
 };
 
-export const markdownBuilder: IBuilder = async (cache: ICache, file: IFile) => {
-  const nameParts = file.name.split(".");
-  nameParts[nameParts.length - 1] = "html";
-
+export const markdownBuilder: IBuilder = async (
+  cache: ICache<IBuilderCache>,
+  file: IFile
+) => {
   const newFile: IFile = {
     ...file,
-    name: nameParts.join("."),
     extension: "html",
     raw: await parseMarkdown(file.raw.toString())
   };
 
-  return htmlBuilder(cache, newFile);
+  return buildHtml(cache, newFile);
 };