+++ src/interfaces/cache.interface.ts
@@ -1,4 +1,0 @@
-export interface ICache {
-  set: (key: string, value: any) => void;
-  get: (key: string) => any;
-}