+++ src/interfaces/index.ts
@@ -1,4 +1,3 @@
 export * from "./builder.interface";
 export * from "./dir-scanner.interface";
-export * from "./cache.interface";
 export * from "./config.interface";