+++ test/unit/builder.metadata.test.ts
@@ -2,28 +2,9 @@ import { getFileMetadata } from "../../src/domain/builder";
 import { IFile } from "../../src/interfaces";
 
 describe("Builder - Metadata", () => {
-  test("Get template metadata (single line)", () => {
+  test("Parse template file", () => {
     const file: IFile = {
-      name: "article.html",
-      extension: "html",
-      modifiedAt: new Date(),
-      path: "",
-      raw:
-        "<!--\n @template main @a 1234 abcd @b abcdef ghijklm @c 12-12-2019 \n-->"
-    };
-
-    const { metadata } = getFileMetadata(file);
-    expect(metadata).toMatchObject({
-      template: "main",
-      a: "1234 abcd",
-      b: "abcdef ghijklm",
-      c: "12-12-2019"
-    });
-  });
-
-  test("Parse template file (multi line)", () => {
-    const file: IFile = {
-      name: "article.html",
+      name: "article",
       extension: "html",
       modifiedAt: new Date(),
       path: "",
@@ -33,20 +14,23 @@ describe("Builder - Metadata", () => {
       @a 1234\n\
       @b abcdef\n\
       @c 12-12-2019\n\
-      -->"
+      @d hi. \"m,y #(n4-m3)!¡?¿ i's Alberto;\n\
+      -->my-content"
     };
-    const { metadata } = getFileMetadata(file);
+    const { metadata, html } = getFileMetadata(file);
     expect(metadata).toMatchObject({
       template: "main",
       a: "1234",
       b: "abcdef",
-      c: "12-12-2019"
+      c: "12-12-2019",
+      d: "hi. \"m,y #(n4-m3)!¡?¿ i's Alberto;"
     });
+    expect(html).toEqual("my-content");
   });
 
   test("Parse template file (multi line spaces)", () => {
     const file: IFile = {
-      name: "article.html",
+      name: "article",
       extension: "html",
       modifiedAt: new Date(),
       path: "",
@@ -56,20 +40,34 @@ describe("Builder - Metadata", () => {
       @a 1234\n\
       @b abcdef\n\
       @c 12-12-2019\n\
-      -->"
+      --> my-content"
     };
-    const { metadata } = getFileMetadata(file);
+    const { metadata, html } = getFileMetadata(file);
     expect(metadata).toMatchObject({
       template: "main",
       a: "1234",
       b: "abcdef",
       c: "12-12-2019"
     });
+    expect(html).toEqual(" my-content");
+  });
+
+  test("Parse template file without metadata", () => {
+    const file: IFile = {
+      name: "article",
+      extension: "html",
+      modifiedAt: new Date(),
+      path: "",
+      raw: "my-content"
+    };
+    const { metadata, html } = getFileMetadata(file);
+    expect(metadata).toMatchObject({});
+    expect(html).toEqual("my-content");
   });
 
   test("Parse template file (error not started)", () => {
     const file: IFile = {
-      name: "article.html",
+      name: "article",
       extension: "html",
       modifiedAt: new Date(),
       path: "",
@@ -86,7 +84,7 @@ describe("Builder - Metadata", () => {
 
   test("Parse template file (error not ended)", () => {
     const file: IFile = {
-      name: "article.html",
+      name: "article",
       extension: "html",
       modifiedAt: new Date(),
       path: "",
@@ -103,7 +101,7 @@ describe("Builder - Metadata", () => {
 
   test("Parse template file (error reversed)", () => {
     const file: IFile = {
-      name: "article.html",
+      name: "article",
       extension: "html",
       modifiedAt: new Date(),
       path: "",