+++ test/unit/dir-scanner.test.ts
@@ -10,8 +10,15 @@ describe("DirScanner", () => {
     const dir = "./test/data/dir-scanner";
     const files = await DirScanner.scanAndGetFiles(dir);
     expect(files).toBeTruthy();
-    expect(files.length).toBe(2);
-    expect(files[0].name).toBe("text.txt");
-    expect(files[1].name).toBe("text2.txt");
+    expect(files.length).toBe(3);
+    expect(files[0].name).toBe("text");
+    expect(files[0].extension).toBe("txt");
+    expect(files[0].path).toBe("");
+    expect(files[2].name).toBe("post");
+    expect(files[2].extension).toBe("md");
+    expect(files[2].path).toBe("subfolder");
+    expect(files[1].name).toBe("executable");
+    expect(files[1].extension).toBe("");
+    expect(files[1].path).toBe("subfolder");
   });
 });