+++ test/unit/loader.test.ts
@@ -1,18 +1,23 @@
 import { Loader } from "../../src/domain/loader";
-import { MemoryCache } from "../../src/lib/cache";
-import { ICache, IConfig, IFile } from "../../src/interfaces";
+import { MemoryCache, ICache } from "../../src/lib/cache";
+import { IConfig, IFile, IBuilderCache } from "../../src/interfaces";
 import { defaultConfig } from "../../src/use-cases/build/config.default";
 
 describe("Loader", () => {
-  let cache: ICache;
+  let cache: ICache<IBuilderCache>;
 
   beforeEach(() => {
-    cache = new MemoryCache();
-    cache.set("config", {
-      ...defaultConfig,
-      projectFolder: "./test/data/project",
-      distFolder: "./test/dist"
-    } as IConfig);
+    const initCache: IBuilderCache = {
+      config: {
+        ...defaultConfig,
+        projectFolder: "./test/data/project",
+        distFolder: "./test/dist"
+      },
+      templates: {},
+      posts: [],
+      snippets: {}
+    };
+    cache = new MemoryCache<IBuilderCache>(initCache);
   });
 
   test("Folder doesn't exist", async () => {
@@ -52,19 +57,41 @@ describe("Loader", () => {
     const loader = new Loader(cache);
     await loader.init();
 
-    expect(Object.keys(cache.get("templates"))).toMatchObject(["main"]);
-    expect(Object.keys(cache.get("snippets"))).toEqual([
-      "analytics",
-      "scripts"
-    ]);
+    // Workaround for CI/CD file modified at date
+    // TODO Test date inside dir-scanner and mock it here
+    const now = new Date();
+    cache.get("templates").main.modifiedAt = now;
+
+    expect(cache.get("templates")).toMatchObject({
+      main: {
+        name: "main",
+        extension: "html",
+        path: "",
+        modifiedAt: now,
+        raw: Buffer.from(
+          "3c6469763e0a20207b7b636f6e74656e747d7d0a3c2f6469763e",
+          "hex"
+        )
+      }
+    });
+
+    expect(cache.get("snippets")).toMatchObject({
+      analytics: {
+        configMode: "prod",
+        value: '<script>\n  const a = "I am an analytic script!";\n</script>'
+      },
+      scripts: {
+        configMode: "all",
+        value: '<script>\n  const a = "I am an script!";\n</script>'
+      }
+    });
+
     expect(
       cache
         .get("posts")
         .every(
           (p: IFile) =>
-            ["first.html", "second.html", "third.html", "fourth.html"].indexOf(
-              p.name
-            ) > -1
+            ["first", "second", "third", "fourth"].indexOf(p.name) > -1
         )
     ).toBeTruthy();
   });