+++ test/unit/memory.cache.test.ts
@@ -1,9 +1,21 @@
-import { MemoryCache } from "../../src/lib/cache";
+import { MemoryCache, ICache } from "../../src/lib/cache";
+
+interface ITest {
+  name: string;
+}
 
 describe("MemoryCache", () => {
+  let cache: ICache<ITest>;
+
+  beforeEach(() => {
+    const initCache: ITest = {
+      name: ""
+    };
+    cache = new MemoryCache<ITest>(initCache);
+  });
+
   test("Set/Get keys", () => {
-    const cache = new MemoryCache();
-    cache.set("test", "hello");
-    expect(cache.get("test")).toBe("hello");
+    cache.set("name", "hii");
+    expect(cache.get("name")).toBe("hii");
   });
 });