+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - Unreleased
+
+## Added
+
+- Recent posts "Read more" link.
+- Replace post metadata inside post.
+- Templates support using another template (one level only).
+
+## Changed
+
+- Updated node to 12.13 LTS.
+- File metadata name doesn't include extension.
+- Templates are stored as a IFile.
+- MemoryCache improved typings.
+
+## Fixed
+
+- Post metada regex now supports special characters.
+
 ## [0.2.0] - 21-10-2019
 
 ### Added